create error navitem and use it in all necessary use cases
