+++ client/homebrew/pages/editPage/editPage.less
@@ -13,10 +13,6 @@
 			cursor : initial;
 			color  : #666;
 		}
-		&.error{
-			position         : relative;
-			background-color : @red;
-		}
 	}
 	.googleDriveStorage {
 		position : relative;
