+++ client/homebrew/pages/homePage/homePage.less
@@ -46,9 +46,5 @@
 		&:hover{
 			background-color: @green;
 		}
-		&.error{
-			position         : relative;
-			background-color : @red;
-		}
 	}
 }
