+++ client/homebrew/pages/newPage/newPage.less
@@ -4,9 +4,5 @@
 		&:hover{
 			background-color: @green;
 		}
-		&.error{
-			position         : relative;
-			background-color : @red;
-		}
 	}
 }
